Fixed fatal bug with MessageHandler debounce checking
The previous method of using debounce checking turned out to be global meaning that users in the same stage will share the same debounce value instead of having their own. This lead to only one user being able to send a message at a given time unless they re-enter the stage to reset the debounce value.